requirements/common.txt
celery
PyJWT
pendulum
Pygments
Markdown
redis
Django
djangorestframework
django-redis
django-filter
diango-cors-headers
psycopg2-binary
phonenumberslite
